lab5/TODO/requirements.txt
asgiref==3.3.1
Django==3.1.7
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
Jinja2==2.11.3
MarkupSafe==1.1.1
psycopg2==2.8.6
PyJWT==1.7.1
pytz==2021.1
sqlparse==0.4.1
